vm/proxy/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
python-dotenv==1.0.1
psutil==5.9.8
pdfplumber==0.11.0
python-Levenshtein==0.25.1
clamd==1.0.2
